fix: contentLineNumber가 Long->Integer로 변경

+++ src/main/java/com/kidchang/lingopress/press/dto/request/TranslateContentLineRequest.java
@@ -2,7 +2,7 @@ package com.kidchang.lingopress.press.dto.request;
 
 public record TranslateContentLineRequest(
     Long pressId,
-    Long contentLineNumber,
+    Integer contentLineNumber,
     String translateText,
     Boolean isCorrect
 ) {
